// src/rooms-sockets/enums/event-emit.ts
export enum EventEmit {
  RECEIVE_MESSAGE = 'receiveMessage',
  UPDATE_AVAILABLE_ROOMS = 'updateAvailableRooms',
  ERROR = 'error',
  TOGGLE_PARTICIPANT_CAMERA = 'toggleParticipantCamera',
  TOGGLE_PARTICIPANT_VOICE = 'toggleParticipantVoice',
  ANSWER = 'answer',
  USER_CONNECTED = 'user-connected',
  USER_DISCONNECTED = 'user-disconnected',
}